Update to version 0.1 - Working development node

+++ nodes/LightSwitch/EncoderSM/EncoderSM.h
@@ -4,5 +4,5 @@
 #include "hardware/pwm.h"
 #include "hardware/rtc.h"
 
-int EncoderSM_init(void);
-int EncoderSM_run(uint *response);
+uint8_t EncoderSM_init(void);
+uint8_t EncoderSM_run(uint *response);
